Added current location marker in map view pt 3
Note this was done in 3 parts because I'm still not 100% comfortable with Intellij's Git plugin

Map view now has a red dot at the player's current location on the map. Also made map region outline thicker for easier view.

Note that currently clicking on the red dot will throw a Null Pointer Exception. This is because it uses the Game's current tile when clicked to generate the currently selected region. Once Tiles and current tile is more fleshed out this should work properly